update np_v_dat_p_np_pp_iobj mask to allow prepositional phrases within it
will be made into a stricter mask once complement phrases are allowed which would break this. grammar is not complete yet.

similar update may also be needed in another case where a NP is in the middle of the construction can be modified,  `np was v_dat_pp_p4 np by np`