recogs_examples_in_rasp.py - support picking up where one left off with `--skip_rows` argument, also report num correct by category so if recombining multiple runs it is easier to combine the scores (otherwise multiply percentage by count but could introduce rounding error)
Google Colab shuts down my runtimes in the middle of a run sometimes, so this is needed to allow it to continue. Note only CPU instance is needed for the RASP program so one can run it on the free tier which has an even higher interrupt rate making the ability to pick it back up potentially more useful.