Tiny updates to README.md to add links (#22)
- add a link for Ahuja et al. 2024 reference so that folks don't need to
open my paper to find it
In justifying that using an "embedding" that encodes the possible parts
of speech and/or verb types for a word is not unrealistic, I write
```
In general, pre-training or using an embedding like GloVe (Pennington et al., 2014) 
could ensure this type of information was available in the embedding;
 when training from scratch for COGS/ReCOGS we expect the
 act of modeling the input sentences to be able to result 
in embeddings with part-of-speech and verb type information, 
to facilitate this one might consider also to adjust the training 
objective as discussed in (Ahuja et al., 2024 ) to explicitly predict 
the input sentences by treating the seq2seq problem as a 
language modeling problem for the input concatenated with output 
instead of as a seq2seq.
```
This adds the link for Ahuja et al. 2024 which mentions advantage of
reformulating seq2seq problems as language modeling problems at
https://direct.mit.edu/tacl/article/doi/10.1162/tacl_a_00733/127877/Learning-Syntax-Without-Planting-Trees
, which connects us to the literature on Transformers with language
modeling objective learning part of speech information into their
embeddings, see (Tenney et al 2019 https://arxiv.org/abs/1905.05950 ).

Btw, the title of Ahuja's paper is "Learning Syntax without Planting
Trees" but I am going further in my RASP-for-ReCOGS paper, I am saying a
tree-structured/hierarchical representation, at least for ReCOGS (and
COGS, see github.com/willy-b/RASP-for-COGS ) is not required at all for
some seemingly "hierarchical" computations, not just that it doesn't
need to be "planted" or have a bias for it forced on the model.

- remove a line break and escaped character that prevented a URL from
being an autolink